requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
appnope==0.1.4
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asttokens==3.0.1
async-lru==2.3.0
attrs==26.1.0
babel==2.18.0
bcrypt==5.0.0
beautifulsoup4==4.14.3
bleach==6.3.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.6
circuitbreaker==2.1.3
click==8.0.4
cloudpickle==3.1.2
comm==0.2.3
cryptography==46.0.5
debugpy==1.8.20
decorator==5.2.1
defusedxml==0.7.1
dnspython==2.8.0
ecdsa==0.19.1
et_xmlfile==2.0.0
executing==2.2.1
fastapi==0.112.0
fastjsonschema==2.21.2
fqdn==1.5.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
ipykernel==7.2.0
ipython==9.10.0
ipython_pygments_lexers==1.1.1
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.3
json5==0.13.0
jsonpointer==3.1.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.8.0
jupyter_core==5.9.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.4
jupyterlab==4.5.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
lark==1.3.1
llvmlite==0.44.0
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
mistune==3.2.0
nbclient==0.10.4
nbconvert==7.17.0
nbformat==5.10.4
nest-asyncio==1.6.0
notebook==7.5.5
notebook_shim==0.2.4
numba==0.61.2
numpy==1.26.4
oci==2.168.2
oci-cli==3.76.1
openpyxl==3.1.5
overrides==7.7.0
packaging==26.0
pandas==3.0.1
pandas-stubs==3.0.0.260204
pandocfilters==1.5.1
parso==0.8.6
passlib==1.7.4
pdfminer.six==20231228
pdfplumber==0.11.4
pexpect==4.9.0
pillow==12.1.1
platformdirs==4.9.4
prometheus_client==0.24.1
prompt-toolkit==3.0.43
psutil==7.2.2
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.3
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pymongo==4.8.0
pyOpenSSL==25.3.0
pypdfium2==5.6.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-json-logger==4.0.0
python-multipart==0.0.9
pytz==2024.2
PyYAML==6.0.2
pyzmq==27.1.0
referencing==0.37.0
requests==2.32.5
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rpds-py==0.30.0
rsa==4.9.1
scikit-learn==1.5.1
scipy==1.17.1
Send2Trash==2.1.0
shap==0.46.0
six==1.17.0
slicer==0.0.8
soupsieve==2.8.3
stack-data==0.6.3
starlette==0.37.2
terminado==0.18.1
terminaltables==3.1.10
threadpoolctl==3.6.0
tinycss2==1.4.0
tornado==6.5.5
tqdm==4.67.3
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uri-template==1.3.0
urllib3==2.6.3
uvicorn==0.30.1
wcwidth==0.6.0
webcolors==25.10.0
webencodings==0.5.1
websocket-client==1.9.0
